only deleting product if clicked on OK

+++ Components/AdminComponents/Products/ProductsTable.js
@@ -17,7 +17,9 @@ const ProductsTable = ({ products }) => {
       buttons: true,
       dangerMode: true,
     }).then((willDelete) => {
-      dispatch(removeProductData(id));
+      if (willDelete) {
+        dispatch(removeProductData(id));
+      }
     });
   };
